+++ NEWS.md
@@ -3,6 +3,9 @@
 ## General
 * Added a `weights` argument in `mixed_model()`. These are simple multipliers on the log-likelihood contributions of each group/cluster, i.e., we presume that there are multiple replicates of each group/cluster denoted by the weights.
 
+* The internal implementation of the `negative.binomial()` family is now considerably faster.
+
+
 # GLMMadaptive 0.6.0
 
 ## General